Develop

+++ src/main/java/codeenthusiast/TrainingCenterApp/security/jwt/JwtUtils.java
@@ -19,7 +19,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${codeenthusiasts.jwtExpirationMs}")
-    private int jwtExpirationMs;
+    private long jwtExpirationMs;
 
     public String generateJwtToken(Authentication authentication) {
